+++ src/ff8/engine.h
@@ -24,4 +24,7 @@
 
 #include "../ff8.h"
 
+#include <sstream>
+
 int ff8_manage_time_engine(int enable_rdtsc);
+std::string ff8_decode_text(const char* encoded_text);